add freezertype meta for prevent fullsync to fastsync
