+++ core/rawdb/schema.go
@@ -81,6 +81,9 @@ var (
 	//stableStateBlockNumberKey tracks the block number for block state that write disk
 	stableStateBlockNumberKey = []byte("StableStateBlockNumber")
 
+	//PruneAncientFlag flag whether prune ancient
+	pruneAncientKey = []byte("PruneAncientFlag")
+
 	// badBlockKey tracks the list of bad blocks seen by local
 	badBlockKey = []byte("InvalidBlock")
